requirements.txt
google-adk
requests
python-dotenv
Flask
google-generativeai